chore(chat): Rename create and client package
