+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["client"]
+members = ["chat"]
 
 [package]
 name = "muchat"
