+++ chat/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "simplex_client"
+name = "chat_client"
 version = "0.1.0"
 edition = "2021"
